Adding a find experts rpc endpoint.
I stick to REST when I can and when it makes sense. The more I learn
about REST, the more I learn not to shoehorn every single thing into it.

We could treat expert like their own full fledged resource, or we could
nest them under /users/. I choose neither and picked the thing that made
the most sense to me. Use the user controller, give expert it's own
namespace in the api, and give it a unique shape (serializer). I think
treating `find_experts` as a first class citizen is important. It is
theheart of the app and I wouldn't like to debilitate it by forcing it
into a rest box that might not be intuitive to integrate with.