Add the expert finder service.
Here I stumbled and went over time. _I AM SURE_ there is a way to do
this in a single sql query in an extremely performant way. If I had
another hour or two to poke around at it, I might get it. Perhaps that
is for a followup branch to be evaluated separately.

The solution I did bang out quickly so that I could spend more time on
the sql is what we have here. I am not happy with it and I would not
ship this to production. It is not a very efficient algorithm, but
O(n+m) isn't the worst.

The method also suffers from readability. 1. We're using recursion, so
it's going to be a little tricky to look at the first time. 2. The
object I'm passing around is a weird hash whose key is the user object
grabbed out of the DB. The seek and found_expert ought to at least be
the same shape. A clear sign this isn't ideal is the format method I had
to tack on at the end of #find_expert to make it resemble a sane return
value.